2.2.6.67
Note: This version might break some of your blacklist configurations.

Added a "whitelist mode" to blacklists.
Every module in the global module blacklist will be in this mode and
adding the module in a certain material's blacklist will re-enable it.
Basically just a XOR between the global module blacklist and the
material module blacklist

Added some more known material names